feat: #WB2-1526, add revision modal to show versions of a page, compare 2 pages or restore one
